Fix duplication in align_ array/parameters and string/character _literal.
Extract common code into new modules.
