+++ lib/rubocop/cop/style/align_array.rb
@@ -6,6 +6,8 @@ module Rubocop
       # Here we check if the elements of a multi-line array literal are
       # aligned.
       class AlignArray < Cop
+        include AutocorrectAlignment
+
         MSG = 'Align the elements of an array literal if they span more ' +
           'than one line.'
 
@@ -18,26 +20,6 @@ module Rubocop
               if current.loc.column != first_element.loc.column
                 convention(current, :expression)
               end
-            end
-          end
-        end
-
-        def autocorrect_action(node)
-          # We can't use the instance variable inside the lambda. That would
-          # just give each lambda the same reference and they would all get
-          # the last value of @column_delta. A local variable fixes the
-          # problem.
-          column_delta = @column_delta
-
-          @corrections << lambda do |corrector|
-            expr = node.loc.expression
-            if column_delta > 0
-              corrector.replace(expr, ' ' * column_delta + expr.source)
-            else
-              range = Parser::Source::Range.new(expr.source_buffer,
-                                                expr.begin_pos + column_delta,
-                                                expr.end_pos)
-              corrector.replace(range, expr.source)
             end
           end
         end
